LPS-113884 Move MappingSelector to common
